// app/src/generated/models/Lppac_launchpadchoicesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Lppac_launchpadchoicesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Lppac_launchpadchoicesstatecode = keyof typeof Lppac_launchpadchoicesstatecode;
export const Lppac_launchpadchoicesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Lppac_launchpadchoicesstatuscode = keyof typeof Lppac_launchpadchoicesstatuscode;

export interface Lppac_launchpadchoicesBase {
  importsequencenumber?: number;
  lppac_choicetype: string;
  lppac_launchpadchoiceid: string;
  lppac_value: string;
  overriddencreatedon?: string;
  statecode: Lppac_launchpadchoicesstatecode;
  statuscode?: Lppac_launchpadchoicesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Lppac_launchpadchoices extends Lppac_launchpadchoicesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}